Fix payment request body amount type

+++ PaytrailSdkExamples/Components/CartSummaryView.swift
@@ -64,8 +64,8 @@ struct CartSummaryView: View {
 struct CartSummaryView_Previews: PreviewProvider {
     static var previews: some View {
         CartSummaryView(items: [
-            ShoppingItem(id: "#1234", productName: "Paytrail Umbrella", description: "", units: 2, price: Int64(15.00), image: "umbrella", currency: "€", upperLimit: 10000),
-            ShoppingItem(id: "#5678", productName: "Paytrail drinking bottle", description: "", units: 1, price: Int64(20.00), image: "bottle", currency: "€", upperLimit: 5000)
+            ShoppingItem(id: "#1234", productName: "Paytrail Umbrella", description: "", units: 2, price: Int(15.00), image: "umbrella", currency: "€", upperLimit: 10000),
+            ShoppingItem(id: "#5678", productName: "Paytrail drinking bottle", description: "", units: 1, price: Int(20.00), image: "bottle", currency: "€", upperLimit: 5000)
         ])
     }
 }
